+++ retirement_api/utils/tests/test_ss_utilities.py
@@ -251,7 +251,7 @@ class UtilitiesTests(unittest.TestCase):
                      ('{0}'.format(self.today.replace(year=self.today.year - 1)), 1),
                      (self.today.replace(year=self.today.year - 20), 20),
                      (self.today.replace(year=self.today.year - 60), 60),
-                     (self.today, None),
+                     (self.today, 0),
                      ('xx', None),
                      (self.today + datetime.timedelta(days=2), None)]
         for pair in age_pairs: