+++ CHANGELOG.md
@@ -6,6 +6,7 @@ We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 - Fix HTML validation errors
 - Removed calculate-age.js (unused)
 - Change button text to the standard typeface (Avenir Next Medium)
+- Added a custom management command for running SSA checks
 - Changed initial interaction for steps 2 and 3
 
 ## 0.4.7